refactor: Add new command top-level CLI command group for 'game'.
